backend/pyproject.toml
[project]
name = "decision-assistant-backend"
version = "0.1.0"
description = "Decision Intelligence Assistant - FastAPI Backend with RAG, ML, and LLM"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
    "tenacity>=9.0.0",
    "loguru>=0.7.0",
    "textblob>=0.18.0",
    "nltk>=3.9.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
# `dev` is used for tests and local notebook work (`uv sync --extra dev`).
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ipykernel>=7.0.0",
    "ipython>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
